requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
flask>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
markdownify>=0.13.0
python-frontmatter>=1.1.0
httpx>=0.27.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0